refactor/#5 [refactor] : user 에서 some 분리
1. user model에 있던 some 관련 로직 분리
2. some-model 생성
3. 명칭 수정
