+++ src/api/routes/main/main.js
@@ -4,6 +4,7 @@ const controller = require('../../../controller/mainController');
 
 module.exports = (app) => {
     app.use('/main', router);
+    
     // 라우터의 get 함수 를 이용해 request url에 대한 업무처리 로직 정의
     router.get('/', function(req, res, next) {
         res.json({state:200});
