+++ src/models/some-query.js
@@ -1,59 +1,9 @@
-let json = {
-    selectUserCardList : 
-    `SELECT 
-        A.user_id
-        , A.nick_name
-        , A.birth
-        , B.location_latitude as latitude
-        , B.location_longitude as longitude
-     FROM user_profile_info A
-        LEFT JOIN user_info B
-        ON A.user_id = B.user_id
-     WHERE NOT EXISTS (
-        SELECT 'X'
-        FROM like_info B
-        WHERE A.user_id = B.other_user_id
-          AND B.user_id = :user_id
-    )
-    AND NOT EXISTS (
-        SELECT 'X'
-        FROM hate_info B
-        WHERE A.user_id = B.other_user_id
-          AND B.user_id = :user_id
-    )
-    AND A.user_id <> :user_id
-    LIMIT 49
-     `,
-    insertHateUser : 
-    `INSERT INTO hate_info 
-        (user_id, other_user_id, registeration_date) 
-    values 
-        (:user_id, other_user_id, NOW())
-    `,
+const json = {
     insertLikeUser :
     `INSERT INTO like_info 
         (user_id, other_user_id, use_yn, registeration_date) 
     values 
         (:user_id,:other_user_id, '1', NOW())
-    `,
-    selectUserAlbum:
-    `SELECT
-        user_id
-        ,image_file_path
-        ,image_file_name
-    FROM user_album_info
-    WHERE user_id IN (
-    `,
-    selectUserProfile:
-    `SELECT
-        user_id
-        ,height
-        ,gender
-        ,job
-        ,fancy
-        ,academy
-    FROM user_profile_info
-    WHERE user_id IN (
     `,
     selectLikeMeList:
     `SELECT
@@ -97,6 +47,13 @@ let json = {
      WHERE A.other_user_id = :user_id
        AND A.user_id = :other_user_id
     `,
+    insertMatchUser:
+    `
+    INSERT INTO match_info
+    (user_id, other_user_id, use_yn, registration_date)
+    VALUES
+    (:user_id, :other_user_id, '1', NOW())
+    `,
     selectMatchUser: 
     `SELECT
         A.user_id
@@ -107,13 +64,6 @@ let json = {
           AND A.to_user_id = B.user_id
      WHERE A.user_id = :userId
        AND A.user_id = B.to_user_id
-    `,
-    insertMatchUser:
-    `
-    INSERT INTO match_info
-    (user_id, other_user_id, use_yn, registration_date)
-    VALUES
-    (:user_id, :other_user_id, '1', NOW())
     `,
     selectMatchList:
     `SELECT
@@ -163,22 +113,7 @@ let json = {
             ON A.user_id = D.user_id
             AND rnum = '1'
      WHERE A.user_id = :user_id
-    `,
-    selectUser:
-    `SELECT 
-        A.user_id
-        ,B.nick_name
-        ,B.birth
-        ,A.location_latitude AS latitude 
-        ,A.location_longitude AS longitude
-    FROM user_info A
-        JOIN user_profile_info B
-        ON A.user_id = B.user_id
-    WHERE A.user_id = :user_id
     `
 }
 
-module.exports = json
-
-
-
+module.exports = json;
