+++ src/service/user-service.js
@@ -1,6 +1,6 @@
 "use strict";
 const database = require('../loaders/database');
-const userStore = require('../models/User');
+const userStore = require('../models/user-model');
 const logger = require('../loaders/logger');
 const profileTitle = {
     "gender":"성별",
@@ -25,7 +25,8 @@ const getUserDetail = async (data) => {
     userStore.setConnectionPool(poolConnection);
     let [rows] = await userStore.selectUser(data);
     rows = rows.length > 0 ? await getUserProfile(rows) : rows;
-    responseObj.result = rows.length > 0 ? await getUserListImages(rows) : rows;  
+    rows = rows.length > 0 ? await getUserListImages(rows) : rows;  
+    responseObj.result = rows[0];
     poolConnection.release();
     return responseObj;
 };
